Readme

+++ src_pf/ft_parse_wstring.c
@@ -6,7 +6,7 @@
 /*   By: kaye <kaye@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/30 21:19:42 by kaye              #+#    #+#             */
-/*   Updated: 2020/11/30 21:56:16 by kaye             ###   ########.fr       */
+/*   Updated: 2020/12/01 23:18:53 by kaye             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,13 +29,13 @@ static int	with_prec(wchar_t *s, int prec)
 int			ft_parse_wstring(wchar_t *s, t_flag flag)
 {
 	int		count;
-	size_t	len;
+	int		len;
 
 	if (!s)
 		s = L"(null)";
 	count = 0;
 	len = ft_strwlen(s);
-	if (flag.prec >= 0 && (size_t)flag.prec > len)
+	if (flag.prec >= 0 && flag.prec > len)
 		flag.prec = len;
 	if (flag.minus)
 		count += with_prec(s, flag.prec);
